AutoVectors.jl Documentation

AutoVectors.jl Documentation

Constructors

Functions

Index

Constructors

AutoVectors.AutoVector — Method

AutoVector(def=0.0,mini::Integer=1,maxi::Integer=0,miniloc::Integer=0)
AutoVector(f::Function,mini::Integer=1,maxi::Integer=0,miniloc::Integer=0)
AutoVector(v::Vector,mini::Integer=1,maxi::Integer=0,miniloc::Integer=0)

Frequently you just use

v = AutoVector(), defaulting to Float64, size 0

or v = AutoVector(0), defaulting to Int64, size 0

An AutoVector expands when written to outside its range. Reading outside its range does not expand the range, and gives def.

Arguments:

def–default element, usually 0.0. Determines the type T of AutoVector{T}

mini and maxi give the index range of the created AutoVector (logical indices, not index in data vector)

miniloc is the location of mini-1 within the data vector, default 0 (data index)

You can initialize an AutoVector with the default, from a function, or by putting in a vector. Most functions and constructors deal with logical indices, Data indices refers to location within the data vector dat, which should not be something to worry about normally.

AutoVectors.makeAutoVectorOfVecs — Method

makeAutoVectorOfVecs(veczero::Vector,mini::Integer,maxi::Integer)

Create an AutoVector that holds Vectors as elements, an AutoVecotr of Vectors. The zero default Vector is veczero.

Functions

AutoVectors.mini — Method

mini(v::AutoVector)

Minimum logical index.

AutoVectors.maxi — Method

maxi(v::AutoVector)

Maximum logical index.

Base.length — Method

length(v::AutoVector)

Logical length.

AutoVectors.arange — Method

avrange(v::AutoVector)

The logical range as mini:maxi. A synonym is arange.

AutoVectors.avvec — Method

avvec(v::AutoVector)

avvec(v) creates a standard Vector with all values from mini to maxi; the new Vector does not know mini; it only knows its length, given by maxi-mini+1.

AutoVectors.avdot — Method

avdot(x::AutoVector,y::AutoVector)

Dot product (with no complex conjugating), sumi xi y_i over the intersection of ranges.

AutoVectors.convolve — Function

convolve(x::AutoVector,y::AutoVector,cut=1.0e-14)		# use absolute cutoff

Convolve with cutoff; no writing of elements if abs(value) < cut

AutoVectors.doprint — Method

doprint([file descriptor],v::AutoVector; spacing = 1)

Print to standard output or a file all the elements.

AutoVectors.fast — Method

fast(v::AutoVector,i)

fast(v,i) is like accessing v[i], but without the check for being outside the logical range. If i is outside, fast may or may not throw a standard range exception depending on whether i lands outside the data vector's range. For use when i is known to be inside the logical range. Combined with @inbounds, there will be no range checking at all, for optimal speed.

AutoVectors.clear! — Method

clear!(v::AutoVector)

Resets v to empty.

AutoVectors.avnorm — Method

avnorm(v::AutoVector)

Norm of AutoVector

AutoVectors.avlocation — Method

avlocation(v::AutoVector,i)

The data location of logical index i.

AutoVectors.avlocmin — Method

avlocmin(v::AutoVector)

The data location of mini(v).

AutoVectors.avlocmax — Method

avlocmax(v::AutoVector)

The data location of maxi(v).

AutoVectors.avtriple — Method

avtriple(x::AutoVector,y::AutoVector,z::AutoVector)

Triple dot product (with no complex conjugating), sumi xi yi zi over the intersection of ranges.

LinearAlgebra.axpy! — Method

axpy!(y::AutoVector,a::Float64,x::AutoVector)	# y += a * x

LinearAlgebra.axpy! — Method

axpy!(y::AutoVector,a::Float64,x::AutoVector, cutoff::Float64)	# y += a * x with cutoff for writing

AutoVectors.avtripconv — Method

avtripconv(u::AutoVector,g::AutoVector,v::AutoVector)

Same as avdot(convolve(u,g),v)

AutoVectors.makeauto — Method

makeauto(v::Vector{Float64},offset::Integer)

Make a new AutoVector out of v, shifting elements to the left by offset. Produces new data vector.

AutoVectors.applyshift — Method

applyshift(x::AutoVector,offset::Integer)

Create a new AutoVector shifted to left by offset, sharing the same data array.

AutoVectors.shrink! — Method

shrink!(x::AutoVector,cut)

Adjust mini and maxi to omit tails where all elements are less than cut in absolute value.

AutoVectors.reverse_ind — Method

reverse_ind(x::AutoVector)

A reflection about the origin; the new AutoVector goes from -maxi to -mini.

Index

AutoVectors.AutoVector

AutoVectors.applyshift

AutoVectors.arange

AutoVectors.avdot

AutoVectors.avlocation

AutoVectors.avlocmax

AutoVectors.avlocmin

AutoVectors.avnorm

AutoVectors.avtripconv

AutoVectors.avtriple

AutoVectors.avvec

AutoVectors.clear!

AutoVectors.convolve

AutoVectors.doprint

AutoVectors.fast

AutoVectors.makeAutoVectorOfVecs

AutoVectors.makeauto

AutoVectors.maxi

AutoVectors.mini

AutoVectors.reverse_ind

AutoVectors.shrink!

Base.length

LinearAlgebra.axpy!

LinearAlgebra.axpy!